Fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40516 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/lib.c
@@ -908,7 +908,7 @@ static uint32_t rand_uint32(void)
      * pseudocode in the paper as:
      * > rotate32((state ^ (state >> 18)) >> 27, state >> 59)
      * Where rotate32() is a clockwise (right) circular rotation. */
-    output_base = (prev_state ^ (prev_state >> 18)) >> 27;
+    output_base = (uint32_t)((prev_state ^ (prev_state >> 18)) >> 27);
     output_rot = prev_state >> 59;
     return (output_base >> output_rot) | (output_base << (-output_rot & 31));
 }
